New Feature: searchEmails
